Various recovery fixes
* Fix job index range for Stateful with batch size > 1
* Fix detection for cases when all jobs in a layer are finished.
  Originally, `value_finished(model, output, cur_end_job_index)` leads
  to out-of-bound reads when `cur_end_job_index == end_job_index`.

Also fixes copying model parameters to NVM. They usualy go beyond 64KB
(follow-up 56b0d26b0ee7a8dde66b99c51bd7027955da0796) and a compiler
warning -Woverflow is reported.

+++ common/platform.cpp
@@ -193,7 +193,7 @@ void first_run(void) {
     my_printf_debug("Init for " CONFIG "/" METHOD " with batch size=%d" NEWLINE, BATCH_SIZE);
 }
 
-void write_to_nvm_segmented(const uint8_t* vm_buffer, uint32_t nvm_offset, uint16_t total_len, uint16_t segment_size) {
+void write_to_nvm_segmented(const uint8_t* vm_buffer, uint32_t nvm_offset, uint32_t total_len, uint16_t segment_size) {
     for (uint16_t idx = 0; idx < total_len; idx += segment_size) {
         write_to_nvm(vm_buffer + idx, nvm_offset + idx, MIN_VAL(total_len - idx, segment_size));
     }
